+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## [Unreleased](https://github.com/TechnologyAdvice/stardust/tree/HEAD)
+
+[Full Changelog](https://github.com/TechnologyAdvice/stardust/compare/v0.16.4...HEAD)
+
+**Closed issues:**
+
+- Compilation issue with docgenInfo.json \(windows\) [\#307](https://github.com/TechnologyAdvice/stardust/issues/307)
+- File docgenInfo.json missing from 'docs/app' [\#304](https://github.com/TechnologyAdvice/stardust/issues/304)
+
+**Merged pull requests:**
+
+- Fix getComponentDocInfo docPath on windows [\#312](https://github.com/TechnologyAdvice/stardust/pull/312) ([levithomason](https://github.com/levithomason))
+
 ## [v0.16.4](https://github.com/TechnologyAdvice/stardust/tree/v0.16.4) (2016-06-28)
 [Full Changelog](https://github.com/TechnologyAdvice/stardust/compare/v0.16.3...v0.16.4)